## Description

<!--- Please give a description of the result including relevant screen shots. -->

## Testing

<!--- Please give a description of any manual testing steps that might be required. -->

## Breaking changes

<!--- Does this PR cause breaking changes that might impact other developers, please describe them here and what other
developers might have to accommodate for. -->

## Checklist
<!---
Please remove any checklist items not relevant to your PR
-->

- [ ] [Synchronization calls](https://dev.azure.com/jibe-erp/JiBe/_workitems/edit/685837) are added (if you added any code to modify the table)
- [ ] This PR fulfills all acceptance criteria
- [ ] Relevant well written tests included

Related work items: #755545, #755549, #755556, #755559, #773437, #773438, #773441, #773443, #773444, #773445, #786689, #786692, #786694

+++ api/src/application-layer/drydock/specification-details/sub-items/DeleteSubItemCommand.ts
@@ -1,22 +1,56 @@
+import { SynchronizerService } from 'j2utils';
+
+import { ApplicationException } from '../../../../bll/drydock/core/exceptions/ApplicationException';
+import { getTableName } from '../../../../common/drydock/ts-helpers/tableName';
 import { validateAgainstModel } from '../../../../common/drydock/ts-helpers/validate-against-model';
-import { DeleteOneParams } from '../../../../dal/drydock/specification-details/sub-items/dto/DeleteOneParams';
+import { SpecificationDetailsRepository } from '../../../../dal/drydock/specification-details/SpecificationDetailsRepository';
+import { DeleteSubItemParams } from '../../../../dal/drydock/specification-details/sub-items/dto/DeleteSubItemParams';
 import { SpecificationDetailsSubItemsRepository } from '../../../../dal/drydock/specification-details/sub-items/SpecificationDetailsSubItemsRepository';
+import { VesselsRepository } from '../../../../dal/drydock/vessels/VesselsRepository';
+import { SpecificationDetailsSubItemEntity } from '../../../../entity/drydock/SpecificationDetailsSubItemEntity';
 import { Command } from '../../core/cqrs/Command';
 import { UnitOfWork } from '../../core/uof/UnitOfWork';
 
-export class DeleteSubItemCommand extends Command<DeleteOneParams, void> {
+export class DeleteSubItemCommand extends Command<DeleteSubItemParams, void> {
     protected readonly subItemsRepo = new SpecificationDetailsSubItemsRepository();
     protected readonly uow = new UnitOfWork();
+    protected readonly tableName = getTableName(SpecificationDetailsSubItemEntity);
+    protected readonly vesselsRepository = new VesselsRepository();
+    protected readonly specificationDetailsRepository = new SpecificationDetailsRepository();
 
-    private params: DeleteOneParams;
-
-    protected async ValidationHandlerAsync(request: DeleteOneParams): Promise<void> {
-        this.params = await validateAgainstModel(DeleteOneParams, request);
-    }
+    private params: DeleteSubItemParams;
 
     public async MainHandlerAsync(): Promise<void> {
+        if (await this.specificationDetailsRepository.isSpecificationIsCompleted(this.params.specificationDetailsUid)) {
+            throw new ApplicationException('Specification is completed, cannot be updated');
+        }
+
         await this.uow.ExecuteAsync(async (queryRunner) => {
+            const vessel = await this.vesselsRepository.GetVesselBySpecification(
+                this.params.specificationDetailsUid,
+                queryRunner,
+            );
+
             await this.subItemsRepo.deleteOneExistingByUid(this.params, queryRunner);
+            await SynchronizerService.dataSynchronizeManager(
+                queryRunner.manager,
+                this.tableName,
+                'uid',
+                this.params.uid,
+                vessel.VesselId,
+            );
+
+            await this.subItemsRepo.deleteAllSubItemRelations(this.params.uid, queryRunner);
+            await SynchronizerService.dataSynchronizeByConditionManager(
+                queryRunner.manager,
+                getTableName(SpecificationDetailsSubItemEntity),
+                vessel.VesselId,
+                `uid = '${this.params.uid}'`,
+            );
         });
+    }
+
+    protected async ValidationHandlerAsync(request: DeleteSubItemParams): Promise<void> {
+        this.params = await validateAgainstModel(DeleteSubItemParams, request);
     }
 }
